Revert "Move temporary code into own module"
This reverts commit ec3188af12dafdc460868b3283dcd40b0def4bd5.
