+++ marconi/marconi.cabal
@@ -126,9 +126,7 @@ test-suite marconi-test
   type:           exitcode-stdio-1.0
   main-is:        Spec.hs
   hs-source-dirs: test
-  other-modules:
-    Gen
-    Integration
+  other-modules:  Integration
 
   --------------------
   -- Local components
